Authentication completed

+++ backend/src/utils/removeFiles.js
@@ -1,6 +1,7 @@
 import fs from "fs";
 
 async function removeFiles(localFilePath) {
+    if(!localFilePath) return console.log("No file to delete");
     await fs.unlinkSync(localFilePath, (err) => {
         if (err) {
             console.error(err);
